inference-pipeline/requirements.txt

google-genai
pillow
numpy
python-dotenv
